Initial back-end stack gems (Gemfile).
This is not a Ruby on Rails application so the following gems were be
carefully selected to create an easy to maintain code that follows the
Clean Architecture and Hexagonal Architecture concepts.

See:

* https://blog.cleancoder.com/uncle-bob/2012/08/13/the-clean-architecture.html
* https://alistair.cockburn.us/hexagonal-architecture/
* https://www.12factor.net/config

dry-system

https://dry-rb.org/gems/dry-system/0.12/

dry-system is an object dependency management system that allows to
configure reusable components in any environment, set up their
load-paths, require needed files and instantiate objects automatically
with the ability of inject them as dependencies.

dry-monads

https://dry-rb.org/gems/dry-monads/1.3/

This gem will be used to implement operations. An operation (a.k.a.
interactor, use-case, or service object) will be responsible to
encapsulate the business logic. The operations are agnostic of gateways
and delivery mechanisms. Instead of that, they depend upon abstractions
not concretions (Dependency inversion principle)

rom, rom-sql and pg

https://rom-rb.org/

Ruby Object Mapper is a persistence and mapping toolkit for Ruby. It
will work as an abstraction between the business logic and the data
layer.

rest-client

https://github.com/rest-client/rest-client

A "simple" HTTP and REST client for Ruby. This gem will be used to
implement the data poller gateway that will fetch the people records
from the SalesLoft REST API.

roda

http://roda.jeremyevans.net/

Roda is a routing tree web toolkit, designed for building fast and
maintainable web applications in ruby. The REST API to expose people
data to the front-end application (delivery mechanism) will be
implemented on the top of this library.

dotenv

https://github.com/bkeepers/dotenv

The app's config will be stored in environment variables. Dotenv gem
allows to easily load environment variables from a .env file.

rake

https://github.com/ruby/rake

Rake is a Make-like program implemented in Ruby. It will allow to
automatize tasks like triggering the logic to import the people records
from SalesLoft API.

rspec

https://rspec.info/

Ruby testing tool that will be use to create unit tests for the different
components of the back-end.

database-cleaner:

http://databasecleaner.github.io/

Allows to clean the database every time the unit test suite runs.

rack-test gem.

https://github.com/rack-test/rack-test

Useful for testing API requests.