Add damerou-levenshtein gem.
This gem contains, among other things, an implementation for Ruby of the
Levenshtein algorithm for calculating the distance between two words. It
will be used to detect whether or not two people are similar based on
the calculated distance for its email addresses.

See: https://en.wikipedia.org/wiki/Levenshtein_distance